chore: clean up duplicate listings systems + standardize URLs
The /invest section had three parallel listing implementations from
different sessions:

1. Old per-vertical client components (FarmlandListingsClient, etc.)
   — already replaced by UnifiedListingsClient but the client files
   were left in the codebase as dead code (~2,000 lines, zero imports)

2. Dynamic /invest/[category]/listings route — built by another session
   but completely shadowed by the literal directories that already exist
   for all 11 categories. Unreachable dead code (295 lines).

3. URL inconsistency — most categories used /listings, but mining used
   /opportunities, renewable-energy used /projects, startups used
   /opportunities. Confusing for users and SEO.

Changes:

Deleted dead code (~2,200 lines):
- 10 unused *ListingsClient.tsx files (Farmland, Business, Commercial,
  Franchise, PrivateCredit, Infrastructure, Alternatives, Mining,
  Energy, Startup)
- /invest/[category]/listings/page.tsx (shadowed by literal routes)

URL standardization:
- /invest/mining/opportunities → /invest/mining/listings
- /invest/renewable-energy/projects → /invest/renewable-energy/listings
- /invest/startups/opportunities → /invest/startups/listings

301 redirects for old paths already exist in next.config.ts so old
links keep working. Updated all internal references in:
- app/page.tsx
- app/invest/[slug]/page.tsx
- app/invest/my-listings/page.tsx
- app/invest/mining/page.tsx
- app/invest/renewable-energy/page.tsx
- app/invest/startups/page.tsx
- app/sitemap.ts
- components/ListingCard.tsx
- components/IntentPicker.tsx

Each per-vertical page keeps its custom hero, breadcrumbs, and SEO
metadata (Option A from the audit) — preserves unique-page SEO benefits
while eliminating tech debt.

TypeScript: clean. Build: succeeds.
