+++ app/invest/[slug]/page.tsx
@@ -68,12 +68,12 @@ export async function generateMetadata({
 
 const LISTINGS_LINKS: Record<string, { href: string; label: string }> = {
   "buy-business": { href: "/invest/buy-business/listings", label: "Browse Businesses for Sale" },
-  mining: { href: "/invest/mining/opportunities", label: "Browse Mining Opportunities" },
+  mining: { href: "/invest/mining/listings", label: "Browse Mining Opportunities" },
   farmland: { href: "/invest/farmland/listings", label: "Browse Farmland Listings" },
   "commercial-property": { href: "/invest/commercial-property/listings", label: "Browse Commercial Properties" },
   franchise: { href: "/invest/franchise/listings", label: "Browse Franchise Opportunities" },
-  "renewable-energy": { href: "/invest/renewable-energy/projects", label: "Browse Energy Projects" },
-  startups: { href: "/invest/startups/opportunities", label: "Browse Startup Opportunities" },
+  "renewable-energy": { href: "/invest/renewable-energy/listings", label: "Browse Energy Projects" },
+  startups: { href: "/invest/startups/listings", label: "Browse Startup Opportunities" },
 };
 
 /* ─────────────────────────────────────────────
