+++ app/invest/mining/page.tsx
@@ -58,7 +58,7 @@ export default function MiningPage() {
           </p>
 
           <Link
-            href="/invest/mining/opportunities"
+            href="/invest/mining/listings"
             className="inline-flex items-center gap-2 bg-amber-500 hover:bg-amber-400 text-slate-900 font-bold text-base px-7 py-3.5 rounded-xl transition-colors shadow-lg"
           >
             Browse Mining Opportunities
@@ -190,7 +190,7 @@ export default function MiningPage() {
             </p>
             <div className="flex flex-col sm:flex-row gap-3 justify-center">
               <Link
-                href="/invest/mining/opportunities"
+                href="/invest/mining/listings"
                 className="inline-flex items-center justify-center gap-2 bg-amber-500 hover:bg-amber-400 text-slate-900 font-bold px-6 py-3 rounded-xl text-sm"
               >
                 Browse Mining Opportunities
