+++ app/sitemap.ts
@@ -28,10 +28,10 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
     "/invest/hybrid-securities", "/invest/crypto-staking", "/invest/smsf",
     "/invest/private-equity", "/invest/bonds", "/invest/gold", "/invest/ipos",
     "/invest/listings",
-    "/invest/buy-business/listings", "/invest/mining/opportunities",
+    "/invest/buy-business/listings", "/invest/mining/listings",
     "/invest/farmland/listings", "/invest/commercial-property/listings",
-    "/invest/franchise/listings", "/invest/renewable-energy/projects",
-    "/invest/startups/opportunities", "/invest/alternatives/listings",
+    "/invest/franchise/listings", "/invest/renewable-energy/listings",
+    "/invest/startups/listings", "/invest/alternatives/listings",
     "/invest/private-credit/listings", "/invest/infrastructure/listings",
     "/foreign-investment/united-states", "/foreign-investment/japan", "/foreign-investment/india",
     "/foreign-investment/malaysia", "/foreign-investment/new-zealand", "/foreign-investment/south-korea",
